tests/wpt/tests/speculation-rules/prerender/navigator-plugins.https.html

<!DOCTYPE html>
<title>Same-origin prerendering can access navigator.plugins</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.js"></script>

<body>
<script>
setup(() => {
  assertSpeculationRulesIsSupported();
  assert_implements_optional(
      'plugins' in navigator, 'navigator.plugins is not provided.'
    );
});

promise_test(async t => {
  const uid = token();
  const bc = new PrerenderChannel('prerender-channel', uid);
  t.add_cleanup(_ => bc.close());

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });

  // Start prerendering a page that attempts to access the navigator.plugins.
  startPrerendering(`resources/navigator-plugins.html?uid=${uid}`);
  const result = await gotMessage;
  const plugins = JSON.parse(result);
  assert_equals(plugins.length, navigator.plugins.length);
  for (let i = 0; i < plugins.length; ++i) {
    const expected_plugin = navigator.plugins[i];
    assert_equals(plugins[i].pluginLength, expected_plugin.length);
    for (let j = 0; j < plugins[i].pluginLength; ++j) {
      assert_equals(plugins[i].pluginTypes[j], expected_plugin[j].type,
                    `type of navigator.plugins[${i}].plugin[${j}]`);
    }
  }
}, 'prerendering page should be able to access the navigator.plugins');
</script>
